requirements.txt
black==23.12.1
mypy==1.19.1
pytest-asyncio